Chapter 8: The Fertility Ritual Update Part 1: Part 3
In this time of pain and devastation, one man rises above all the rest as a beacon of hope in a dying land. That man is me, because I finally managed to fucking get this shit into a playable state.
There are still bugs and tons of things I'd like to fix but frankly most of that has more to do with HAR than me, so I figure I'll just save it all for Part 2 when hopefully things are more settled.
Why am I writing this, I'm literally the only person that uses this mod except the evil one, and I'm 99.99% sure nobody is reading this. If you are reading it, leave a comment below. Why not? But you won't, because you don't exist. Bla bla bla... can I just play the damn game now?
Please send a prayer to Allah/God/Yawheh/Guanyin/Whichever Hindu god you pray to for this sort of situation/Some other guy/idk.