Chapter 30: The Magic Must Die, so That the Mod Can Live
After eons of profound introspection and soul-wrenching deliberation, the architects of destiny have come to a conclusion that shall forever alter the course of history.

With heavy hearts and trembling hands, we stand on the precipice of a new era, a terrifying leap into the unknown. The time has come to sever the ties that bind us to the enchanted realm of A Rimworld of Magic. This unholy union, once thought to be an unbreakable bond, must now be rent asunder, like the veil between worlds torn by the claws of fate.

Oh, the anguish that consumes us as we contemplate this monumental schism!